backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pillow
httpx==0.25.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-cloud-vision==3.4.4
google-generativeai
google-auth==2.40.0
google-search-results==2.4.2

# SerpAPI逆画像検索用
imagehash>=4.3.1
requests>=2.31.0

# PDF処理用ライブラリ（推奨：PyMuPDF）
PyMuPDF>=1.23.0

# 代替案（PyMuPDFが利用できない場合）
# pdf2image>=1.16.0
# PyPDF2>=3.0.0